+++ CHANGELOG.md
@@ -6,15 +6,16 @@ in 2.x, 1.x and 0.x versions.
 
 * 2.1.2 (WIP)
 
+  * Fixed display units from Preferences when using fractional inches
   * Fixed bug when preset's name contains numbers only
   * Fixed text input reset value
   * Added retina screen support for highlight part tool
-  
+
 * 2.1.1 (2021-08-02)
 
   * Added thickness fields to labels
   * Added total cut length in cutting diagram summary
-  * Added model description in printable headers if it's defined
+  * Added model description in printable headers if it is defined
   * Fixed forum conversation tag filter
   * Fixed bug in preset when names contains quote
   * Improved dimensions display for small parts in cutting diagrams